refactor: define tool with server.tool
